requirements.txt
# SSZ Projection Suite - Python Dependencies
#
# Copyright © 2025
# Carmen Wrede und Lino Casu
# Licensed under the ANTI-CAPITALIST SOFTWARE LICENSE v1.4
#
# Core scientific stack
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0
matplotlib>=3.7.0
sympy>=1.12

# Astronomy and physics
astropy>=5.3.0
astroquery>=0.4.6

# Configuration and I/O
pyyaml>=6.0

# Testing framework
pytest>=8.0.0
pytest-timeout>=2.1.0
colorama>=0.4.6
iniconfig>=2.0.0
exceptiongroup>=1.1.0; python_version < "3.11"

# Visualization (optional but recommended)
plotly>=5.18.0
narwhals>=1.15.1
kaleido>=0.2.1

# Data validation (optional)
pydantic>=2.0.0

# Parquet support for GAIA data
pyarrow>=14.0.0

# Jupyter support (optional)
jupyter>=1.0.0
ipykernel>=6.25.0
